webapp/backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
yfinance>=0.2.36
TA-Lib>=0.4.28
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
